+++ src/main/java/org/cotato/csquiz/domain/auth/service/EmailVerificationService.java
@@ -5,19 +5,13 @@ import static org.cotato.csquiz.common.util.EmailUtil.createSignupApprovedMessag
 import static org.cotato.csquiz.common.util.EmailUtil.createSignupRejectionMessageBody;
 import static org.cotato.csquiz.common.util.EmailUtil.getVerificationMessageBody;
 import static org.cotato.csquiz.domain.auth.constant.EmailConstants.CONVERSION_TO_OM_SUBJECT;
-import static org.cotato.csquiz.domain.auth.constant.EmailConstants.SENDER_EMAIL;
-import static org.cotato.csquiz.domain.auth.constant.EmailConstants.SENDER_PERSONAL;
 import static org.cotato.csquiz.domain.auth.constant.EmailConstants.SIGNUP_REJECT_SUBJECT;
 import static org.cotato.csquiz.domain.auth.constant.EmailConstants.SIGNUP_SUCCESS_SUBJECT;
 
-import jakarta.mail.Message.RecipientType;
-import jakarta.mail.MessagingException;
-import jakarta.mail.internet.InternetAddress;
-import jakarta.mail.internet.MimeMessage;
-import java.io.UnsupportedEncodingException;
 import java.util.concurrent.ThreadLocalRandom;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.cotato.csquiz.common.email.EmailService;
 import org.cotato.csquiz.common.error.exception.AppException;
 import org.cotato.csquiz.common.error.ErrorCode;
 import org.cotato.csquiz.domain.auth.cache.EmailRedisRepository;
@@ -25,7 +19,6 @@ import org.cotato.csquiz.domain.auth.entity.Member;
 import org.cotato.csquiz.domain.auth.enums.EmailType;
 import org.cotato.csquiz.domain.auth.utils.EmailFormValidator;
 import org.cotato.csquiz.domain.auth.cache.VerificationCodeRedisRepository;
-import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -38,7 +31,7 @@ public class EmailVerificationService {
     private static final int CODE_LENGTH = 6;
     private static final int CODE_BOUNDARY = 10;
 
-    private final JavaMailSender mailSender;
+    private final EmailService emailService;
     private final VerificationCodeRedisRepository verificationCodeRedisRepository;
     private final EmailFormValidator emailFormValidator;
     private final EmailRedisRepository emailRedisRepository;
@@ -54,7 +47,7 @@ public class EmailVerificationService {
 
         String verificationMessage = getVerificationMessageBody(verificationCode);
 
-        sendEmail(recipient, verificationMessage, subject);
+        emailService.sendEmail(recipient, verificationMessage, subject);
     }
 
     public void sendSignUpApprovedToEmail(Member recipientMember) {
@@ -62,7 +55,7 @@ public class EmailVerificationService {
 
         String successMessage = createSignupApprovedMessageBody(recipientMember);
 
-        sendEmail(recipientMember.getEmail(), successMessage, SIGNUP_SUCCESS_SUBJECT);
+        emailService.sendEmail(recipientMember.getEmail(), successMessage, SIGNUP_SUCCESS_SUBJECT);
         log.info("가입 승인 완료 이메일 전송 완료");
     }
 
@@ -71,7 +64,7 @@ public class EmailVerificationService {
 
         String rejectMessage = createSignupRejectionMessageBody(recipientMember);
 
-        sendEmail(recipientMember.getEmail(), rejectMessage, SIGNUP_REJECT_SUBJECT);
+        emailService.sendEmail(recipientMember.getEmail(), rejectMessage, SIGNUP_REJECT_SUBJECT);
         log.info("가입 승인 거절 이메일 전송 완료");
     }
 
@@ -80,7 +73,7 @@ public class EmailVerificationService {
 
         String conversionMessageBody = createOldMemberConversionEmailBody(recipientMember);
 
-        sendEmail(recipientMember.getEmail(), conversionMessageBody, CONVERSION_TO_OM_SUBJECT);
+        emailService.sendEmail(recipientMember.getEmail(), conversionMessageBody, CONVERSION_TO_OM_SUBJECT);
         log.info("OM 전환 이메일 전송 완료");
     }
 
@@ -93,28 +86,6 @@ public class EmailVerificationService {
         return String.valueOf(builder);
     }
 
-    private void sendEmail(String recipient, String messageBody, String subject) {
-        try {
-            MimeMessage message = mailSender.createMimeMessage();
-
-            message.addRecipients(RecipientType.TO, recipient);
-            message.setSubject(subject);
-            message.setText(messageBody, "utf-8", "html");
-            message.setFrom(getInternetAddress());
-            mailSender.send(message);
-            log.info("이메일 전송 완료");
-        } catch (MessagingException e) {
-            throw new AppException(ErrorCode.EMAIL_SEND_ERROR);
-        }
-    }
-
-    private InternetAddress getInternetAddress() {
-        try {
-            return new InternetAddress(SENDER_EMAIL, SENDER_PERSONAL);
-        } catch (UnsupportedEncodingException e) {
-            throw new RuntimeException(e);
-        }
-    }
 
     public void verifyCode(EmailType type, String email, String code) {
         String savedVerificationCode = verificationCodeRedisRepository.getByEmail(type, email);
